Separate classes from Noah class
