+++ src/main/java/NoahException.java
@@ -1,0 +1,5 @@
+public class NoahException extends Exception {
+    public NoahException(String message) {
+        super(message);
+    }
+}
